Refactor x509 cert into focused modules
