+++ crates/uselesskey-x509/src/cert.rs
@@ -1,32 +1,17 @@
 //! X.509 certificate generation and output.
 
+mod material;
+mod negative;
+mod output;
+
 use std::fmt;
 use std::sync::Arc;
 
-use rand_chacha::ChaCha20Rng;
-use rand_core::RngCore;
-use rand_core::SeedableRng;
-use rcgen::{
-    BasicConstraints, CertificateParams, DnType, ExtendedKeyUsagePurpose, IsCa, KeyPair,
-    KeyUsagePurpose, PKCS_RSA_SHA256,
-};
-use rustls_pki_types::PrivatePkcs8KeyDer;
-use time::Duration as TimeDuration;
-use uselesskey_core::negative::CorruptPem;
-use uselesskey_core::sink::TempArtifact;
-use uselesskey_core::{Error, Factory};
-use uselesskey_rsa::{RsaFactoryExt, RsaSpec};
+use uselesskey_core::Factory;
 
+use self::material::Inner;
 use crate::chain::X509Chain;
-use crate::negative::{
-    corrupt_cert_der_deterministic, corrupt_cert_pem, corrupt_cert_pem_deterministic,
-    truncate_cert_der,
-};
-use crate::srp::derive::{
-    deterministic_base_time_from_parts, deterministic_serial_number_with_rng,
-};
-use crate::srp::negative::X509Negative;
-use crate::srp::spec::{ChainSpec, NotBeforeOffset, X509Spec};
+use crate::srp::spec::{ChainSpec, X509Spec};
 
 /// Cache domain for X.509 certificate fixtures.
 ///
@@ -58,13 +43,6 @@ pub struct X509Cert {
     label: String,
     spec: X509Spec,
     inner: Arc<Inner>,
-}
-
-struct Inner {
-    cert_der: Arc<[u8]>,
-    cert_pem: String,
-    private_key_pkcs8_der: Arc<[u8]>,
-    private_key_pkcs8_pem: String,
 }
 
 impl fmt::Debug for X509Cert {
@@ -122,475 +100,13 @@ impl X509FactoryExt for Factory {
     }
 }
 
-impl X509Cert {
-    fn new(factory: Factory, label: &str, spec: X509Spec) -> Self {
-        let inner = load_inner(&factory, label, &spec, "good");
-        Self {
-            factory,
-            label: label.to_string(),
-            spec,
-            inner,
-        }
-    }
-
-    #[allow(
-        dead_code,
-        reason = "reserved for future variant-based negative fixtures"
-    )]
-    fn load_variant(&self, variant: &str) -> Arc<Inner> {
-        load_inner(&self.factory, &self.label, &self.spec, variant)
-    }
-
-    // =========================================================================
-    // Certificate outputs
-    // =========================================================================
-
-    /// DER-encoded certificate bytes.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// assert!(!cert.cert_der().is_empty());
-    /// ```
-    pub fn cert_der(&self) -> &[u8] {
-        &self.inner.cert_der
-    }
-
-    /// PEM-encoded certificate.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// assert!(cert.cert_pem().starts_with("-----BEGIN CERTIFICATE-----"));
-    /// ```
-    pub fn cert_pem(&self) -> &str {
-        &self.inner.cert_pem
-    }
-
-    /// DER-encoded PKCS#8 private key bytes.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// assert!(!cert.private_key_pkcs8_der().is_empty());
-    /// ```
-    pub fn private_key_pkcs8_der(&self) -> &[u8] {
-        &self.inner.private_key_pkcs8_der
-    }
-
-    /// PEM-encoded PKCS#8 private key.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// assert!(cert.private_key_pkcs8_pem().contains("-----BEGIN PRIVATE KEY-----"));
-    /// ```
-    pub fn private_key_pkcs8_pem(&self) -> &str {
-        &self.inner.private_key_pkcs8_pem
-    }
-
-    /// Combined PEM containing both certificate and private key.
-    ///
-    /// This is a common format for TLS server configuration where
-    /// a single file holds the server identity (cert + key).
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// let identity = cert.identity_pem();
-    /// assert!(identity.contains("-----BEGIN CERTIFICATE-----"));
-    /// assert!(identity.contains("-----BEGIN PRIVATE KEY-----"));
-    /// ```
-    pub fn identity_pem(&self) -> String {
-        format!("{}\n{}", self.cert_pem(), self.private_key_pkcs8_pem())
-    }
-
-    // =========================================================================
-    // Tempfile outputs
-    // =========================================================================
-
-    /// Write the PEM certificate to a tempfile.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// let temp = cert.write_cert_pem().unwrap();
-    /// assert!(temp.path().exists());
-    /// ```
-    pub fn write_cert_pem(&self) -> Result<TempArtifact, Error> {
-        TempArtifact::new_string("uselesskey-", ".crt.pem", self.cert_pem())
-    }
-
-    /// Write the DER certificate to a tempfile.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// let temp = cert.write_cert_der().unwrap();
-    /// assert!(temp.path().exists());
-    /// ```
-    pub fn write_cert_der(&self) -> Result<TempArtifact, Error> {
-        TempArtifact::new_bytes("uselesskey-", ".crt.der", self.cert_der())
-    }
-
-    /// Write the PEM private key to a tempfile.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// let temp = cert.write_private_key_pem().unwrap();
-    /// assert!(temp.path().exists());
-    /// ```
-    pub fn write_private_key_pem(&self) -> Result<TempArtifact, Error> {
-        TempArtifact::new_string("uselesskey-", ".key.pem", self.private_key_pkcs8_pem())
-    }
-
-    /// Write the combined identity PEM (cert + key) to a tempfile.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// let temp = cert.write_identity_pem().unwrap();
-    /// assert!(temp.path().exists());
-    /// ```
-    pub fn write_identity_pem(&self) -> Result<TempArtifact, Error> {
-        TempArtifact::new_string("uselesskey-", ".identity.pem", &self.identity_pem())
-    }
-
-    // =========================================================================
-    // Negative fixtures
-    // =========================================================================
-
-    /// Produce a corrupted variant of the certificate PEM.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_core::negative::CorruptPem;
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// let bad = cert.corrupt_cert_pem(CorruptPem::BadHeader);
-    /// assert!(bad.contains("CORRUPTED"));
-    /// ```
-    pub fn corrupt_cert_pem(&self, how: CorruptPem) -> String {
-        corrupt_cert_pem(self.cert_pem(), how)
-    }
-
-    /// Produce a deterministic corrupted certificate PEM using a variant string.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// let bad = cert.corrupt_cert_pem_deterministic("corrupt:v1");
-    /// assert!(!bad.is_empty());
-    /// ```
-    pub fn corrupt_cert_pem_deterministic(&self, variant: &str) -> String {
-        corrupt_cert_pem_deterministic(self.cert_pem(), variant)
-    }
-
-    /// Produce a truncated variant of the certificate DER.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// let truncated = cert.truncate_cert_der(10);
-    /// assert_eq!(truncated.len(), 10);
-    /// ```
-    pub fn truncate_cert_der(&self, len: usize) -> Vec<u8> {
-        truncate_cert_der(self.cert_der(), len)
-    }
-
-    /// Produce a deterministic corrupted certificate DER using a variant string.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// let bad = cert.corrupt_cert_der_deterministic("corrupt:v1");
-    /// assert!(!bad.is_empty());
-    /// ```
-    pub fn corrupt_cert_der_deterministic(&self, variant: &str) -> Vec<u8> {
-        corrupt_cert_der_deterministic(self.cert_der(), variant)
-    }
-
-    /// Generate a negative fixture variant of this certificate.
-    ///
-    /// The variant is cached separately from the valid certificate.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec, X509Negative};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// let expired = cert.negative(X509Negative::Expired);
-    /// assert_ne!(cert.cert_der(), expired.cert_der());
-    /// ```
-    pub fn negative(&self, negative_type: X509Negative) -> X509Cert {
-        let modified_spec = negative_type.apply_to_spec(&self.spec);
-        let variant = negative_type.variant_name();
-        let inner = load_inner_with_spec(&self.factory, &self.label, &modified_spec, variant);
-
-        X509Cert {
-            factory: self.factory.clone(),
-            label: self.label.clone(),
-            spec: modified_spec,
-            inner,
-        }
-    }
-
-    /// Get a certificate that is already expired.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// let expired = cert.expired();
-    /// assert_ne!(cert.cert_der(), expired.cert_der());
-    /// ```
-    pub fn expired(&self) -> X509Cert {
-        self.negative(X509Negative::Expired)
-    }
-
-    /// Get a certificate that is not yet valid.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// let future = cert.not_yet_valid();
-    /// assert_ne!(cert.cert_der(), future.cert_der());
-    /// ```
-    pub fn not_yet_valid(&self) -> X509Cert {
-        self.negative(X509Negative::NotYetValid)
-    }
-
-    /// Get a certificate with wrong key usage flags.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("svc", X509Spec::self_signed("svc.example.com"));
-    /// let wrong = cert.wrong_key_usage();
-    /// assert!(wrong.spec().is_ca);
-    /// ```
-    pub fn wrong_key_usage(&self) -> X509Cert {
-        self.negative(X509Negative::WrongKeyUsage)
-    }
-
-    // =========================================================================
-    // Metadata
-    // =========================================================================
-
-    /// Get the specification used to create this certificate.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let spec = X509Spec::self_signed("svc.example.com");
-    /// let cert = fx.x509_self_signed("svc", spec.clone());
-    /// assert_eq!(cert.spec().subject_cn, "svc.example.com");
-    /// ```
-    pub fn spec(&self) -> &X509Spec {
-        &self.spec
-    }
-
-    /// Get the label used to create this certificate.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use uselesskey_core::{Factory, Seed};
-    /// # use uselesskey_x509::{X509FactoryExt, X509Spec};
-    /// let fx = Factory::deterministic(Seed::from_env_value("test-seed").unwrap());
-    /// let cert = fx.x509_self_signed("my-svc", X509Spec::self_signed("svc.example.com"));
-    /// assert_eq!(cert.label(), "my-svc");
-    /// ```
-    pub fn label(&self) -> &str {
-        &self.label
-    }
-}
-
-fn load_inner(factory: &Factory, label: &str, spec: &X509Spec, variant: &str) -> Arc<Inner> {
-    load_inner_with_spec(factory, label, spec, variant)
-}
-
-fn load_inner_with_spec(
-    factory: &Factory,
-    label: &str,
-    spec: &X509Spec,
-    variant: &str,
-) -> Arc<Inner> {
-    let spec_bytes = spec.stable_bytes();
-
-    factory.get_or_init(DOMAIN_X509_CERT, label, &spec_bytes, variant, |seed| {
-        let mut rng = ChaCha20Rng::from_seed(*seed.bytes());
-        // Generate RSA key using uselesskey-rsa for deterministic key generation.
-        // We use the label + variant to derive a unique key.
-        let key_label = format!("{}-key", label);
-        let rsa_spec = RsaSpec::new(spec.rsa_bits);
-        let rsa_keypair = factory.rsa(&key_label, rsa_spec);
-
-        // Get the PKCS#8 DER key and convert it to rcgen's KeyPair
-        let pkcs8_der = rsa_keypair.private_key_pkcs8_der();
-        let pkcs8_key = PrivatePkcs8KeyDer::from(pkcs8_der.to_vec());
-        let key_pair =
-            KeyPair::from_pkcs8_der_and_sign_algo(&pkcs8_key, &PKCS_RSA_SHA256).expect("key parse");
-
-        // Build certificate parameters
-        let mut params = CertificateParams::default();
-        params
-            .distinguished_name
-            .push(DnType::CommonName, spec.subject_cn.clone());
-
-        // Set validity period based on spec
-        let rsa_bits = (spec.rsa_bits as u32).to_be_bytes();
-        let base_time = deterministic_base_time_from_parts(&[
-            label.as_bytes(),
-            spec.subject_cn.as_bytes(),
-            spec.issuer_cn.as_bytes(),
-            &rsa_bits,
-        ]);
-
-        let not_before = match spec.not_before_offset {
-            NotBeforeOffset::DaysAgo(days) => base_time - TimeDuration::days(days as i64),
-            NotBeforeOffset::DaysFromNow(days) => base_time + TimeDuration::days(days as i64),
-        };
-
-        let not_after = not_before + TimeDuration::days(spec.validity_days as i64);
-
-        params.not_before = not_before;
-        params.not_after = not_after;
-        params.serial_number = Some(deterministic_serial_number_with_rng(|bytes| {
-            rng.fill_bytes(bytes);
-        }));
-
-        // Set CA status
-        if spec.is_ca {
-            params.is_ca = IsCa::Ca(BasicConstraints::Unconstrained);
-        } else {
-            params.is_ca = IsCa::NoCa;
-        }
-
-        // Set key usage
-        let mut key_usages = Vec::new();
-        if spec.key_usage.digital_signature {
-            key_usages.push(KeyUsagePurpose::DigitalSignature);
-        }
-        if spec.key_usage.key_encipherment {
-            key_usages.push(KeyUsagePurpose::KeyEncipherment);
-        }
-        if spec.key_usage.key_cert_sign {
-            key_usages.push(KeyUsagePurpose::KeyCertSign);
-        }
-        if spec.key_usage.crl_sign {
-            key_usages.push(KeyUsagePurpose::CrlSign);
-        }
-        params.key_usages = key_usages;
-
-        // Add extended key usage for TLS
-        if !spec.is_ca {
-            params.extended_key_usages = vec![
-                ExtendedKeyUsagePurpose::ServerAuth,
-                ExtendedKeyUsagePurpose::ClientAuth,
-            ];
-        }
-
-        // Add Subject Alternative Names
-        let mut sorted_sans = spec.sans.clone();
-        sorted_sans.sort();
-        sorted_sans.dedup();
-        for san in &sorted_sans {
-            params.subject_alt_names.push(rcgen::SanType::DnsName(
-                san.clone().try_into().expect("valid DNS name"),
-            ));
-        }
-
-        // Generate the self-signed certificate
-        let cert = params.self_signed(&key_pair).expect("cert generation");
-
-        let cert_der: Arc<[u8]> = Arc::from(cert.der().as_ref());
-        let cert_pem = cert.pem();
-
-        let private_key_pkcs8_der: Arc<[u8]> = Arc::from(pkcs8_der);
-        let private_key_pkcs8_pem = rsa_keypair.private_key_pkcs8_pem().to_string();
-
-        Inner {
-            cert_der,
-            cert_pem,
-            private_key_pkcs8_der,
-            private_key_pkcs8_pem,
-        }
-    })
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::srp::spec::NotBeforeOffset;
     use crate::testutil::fx;
     use uselesskey_core::Seed;
+    use uselesskey_core::negative::CorruptPem;
 
     #[test]
     fn test_self_signed_cert_generation() {
